Align integer sizes to address G115

+++ controllers/propagator/replicatedpolicy_setup.go
@@ -20,7 +20,7 @@ import (
 
 func (r *ReplicatedPolicyReconciler) SetupWithManager(
 	mgr ctrl.Manager,
-	maxConcurrentReconciles uint,
+	maxConcurrentReconciles uint16,
 	dependenciesSource source.Source,
 	updateSrc source.Source,
 	templateSrc source.Source,
